requirements.txt
flask==3.0.0
psutil==5.9.6
pytest==7.4.0
flask-basicauth==0.2.0
flask-limiter==3.5.0
flask-talisman==1.1.0
python-dotenv==1.0.0
